backend/requirements.txt

Flask
flask-cors
torch
torchvision
pillow
timm
